maestro clientes

+++ BE/BEProducto.cs
@@ -8,7 +8,7 @@ namespace BE
 {
     public class BEProducto
     {
-        public int CodigoProducto{ get; set; }
+        public long CodigoProducto{ get; set; }
         public string Modelo { get; set; }
         public string Descripcion { get; set; }
         public string Marca { get; set;}
@@ -17,7 +17,7 @@ namespace BE
         public int Stock { get; set; }
         public int Almacenamiento { get; set; }
 
-        public BEProducto(int codigoProducto, string modelo, string descripcion, string marca, string color, double precio, int stock, int almacenamiento) 
+        public BEProducto(long codigoProducto, string modelo, string descripcion, string marca, string color, double precio, int stock, int almacenamiento) 
         {
             this.CodigoProducto = codigoProducto;
             this.Modelo = modelo;
